Move timelines post selection methods to the separate DBAdapter trait
