+++ app/support/DbAdapter/posts.js
@@ -2,11 +2,9 @@ import _ from 'lodash';
 import validator from 'validator'
 import pgFormat from 'pg-format';
 
-import { Post, Comment } from '../../models';
+import { Post } from '../../models';
 
-import { initObject, prepareModelPayload, unexistedUID } from './utils';
-import { COMMENT_FIELDS, initCommentObject } from './comments';
-import { ATTACHMENT_FIELDS, initAttachmentObject } from './attachments';
+import { initObject, prepareModelPayload } from './utils';
 
 ///////////////////////////////////////////////////
 // Posts
@@ -237,230 +235,6 @@ const postsTrait = (superClass) => class extends superClass {
     return result;
   }
 
-  /**
-   * Returns UIDs of timelines posts
-   *
-   * @param {number[]|null} timelineIntIds null means select everything
-   * @param {string|null} viewerId UID of the authenticated viewer
-   * @param {object} params
-   * @returns {string[]}
-   */
-  async getTimelinePostsIds(timelineIntIds = null, viewerId = null, params = {}) {
-    params = {
-      limit:           30,
-      offset:          0,
-      sort:            'bumped',
-      withLocalBumps:  false,
-      withoutDirects:  false,
-      createdBefore:   null,
-      createdAfter:    null,
-      activityFeedIds: [],
-      authorsIds:      [],
-      ...params,
-    };
-
-    params.withLocalBumps = params.withLocalBumps && !!viewerId && params.sort === 'bumped';
-
-    // Private feeds viewer can read
-    let visiblePrivateFeedIntIds = [];
-    // Users who banned viewer or banned by viewer (viewer should not see their posts)
-    let  bannedUsersIds = [];
-    // Additional condition for params.withoutDirects option
-    let noDirectsSQL = 'true';
-    let postsAuthorsSQL = null;
-
-    if (viewerId) {
-      [
-        visiblePrivateFeedIntIds,
-        bannedUsersIds,
-      ] = await Promise.all([
-        this.getVisiblePrivateFeedIntIds(viewerId),
-        this.getUsersBansOrWasBannedBy(viewerId),
-      ]);
-
-      if (params.withoutDirects) {
-        // Do not show directs-only messages (any messages posted to the viewer's 'Directs' feed and to ONE other feed)
-        const [directsIntId] = await this.database.pluck('id').from('feeds').where({ user_id: viewerId, name: 'Directs' });
-        noDirectsSQL = `not (destination_feed_ids && '{${directsIntId}}' and array_length(destination_feed_ids, 1) = 2)`;
-      }
-
-      if (params.authorsIds.length > 0) {
-        // Also show posts from these authors
-        postsAuthorsSQL = pgFormat('p.user_id in (%L)', params.authorsIds);
-      }
-    }
-
-    let sourceConditionSQL = 'true'; // select everything
-
-    if (timelineIntIds) {
-      const sourceConditionParts = [];
-      sourceConditionParts.push(pgFormat('p.feed_ids && %L', `{${timelineIntIds.join(',')}}`));
-
-      if (params.activityFeedIds.length > 0) {
-        sourceConditionParts.push(pgFormat('p.feed_ids && %L and p.is_propagable', `{${params.activityFeedIds.join(',')}}`));
-      }
-
-      if (postsAuthorsSQL) {
-        sourceConditionParts.push(postsAuthorsSQL);
-      }
-
-      sourceConditionSQL = `(${sourceConditionParts.join(' or ')})`;
-    }
-
-    const createdAtParts = [];
-
-    if (params.createdBefore) {
-      createdAtParts.push(pgFormat('p.created_at < %L', params.createdBefore));
-    }
-
-    if (params.createdAfter) {
-      createdAtParts.push(pgFormat('p.created_at > %L', params.createdAfter));
-    }
-
-    const createdAtSQL = createdAtParts.length === 0 ? 'true' : createdAtParts.join(' and ');
-    const privacyCondition = viewerId ?
-      pgFormat(`(not p.is_private or p.destination_feed_ids && %L)`, `{${visiblePrivateFeedIntIds.join(',')}}`)
-      : 'not p.is_protected';
-    const bansSQL = bannedUsersIds.length > 0 ?
-      pgFormat(`(not p.user_id in (%L))`, bannedUsersIds)
-      : 'true';
-
-    const restrictionsSQL = [bansSQL, privacyCondition, noDirectsSQL, createdAtSQL].join(' and ');
-
-    const maxOffsetWithLocalBumps = 1000;
-    const smallFeedThreshold = 5;
-
-    /**
-     * PostgreSQL is not very good dealing with queries like
-     * `select ... from ... where COND order by ORD limit LIM`
-     * when COND selects only a small amount of all records. See
-     * for example https://stackoverflow.com/a/6038853.
-     *
-     * So we using a following heuristics here: if feed consists of
-     * few (<= 5) source timelines then CTE is used, otherwise
-     * normal 'where' is used.
-     *
-     * @param {number} limit
-     * @param {number} offset
-     * @param {string} sort
-     */
-    const getPostsSQL = (limit, offset, sort) => {
-      if (timelineIntIds && timelineIntIds.length <= smallFeedThreshold) {
-        // Request with CTE for the relatively small feed
-        return pgFormat(`
-          with posts as (
-            select * from posts p where ${sourceConditionSQL}
-          )
-          select p.uid, p.bumped_at as date
-          from 
-            posts p
-          where
-            ${restrictionsSQL}
-          order by
-            p.%I desc
-          limit %L offset %L
-        `, `${sort}_at`, limit, offset);
-      }
-
-      // Request without CTE for the large (tipically RiverOfNews) feed
-      return pgFormat(`
-        select p.uid, p.bumped_at as date
-        from 
-          posts p
-        where
-          ${sourceConditionSQL} and ${restrictionsSQL}
-        order by
-          p.%I desc
-        limit %L offset %L
-      `, `${sort}_at`, limit, offset);
-    };
-
-    if (!params.withLocalBumps || params.offset > maxOffsetWithLocalBumps) {
-      // without local bumps
-      const sql = getPostsSQL(params.limit, params.offset, params.sort);
-      return (await this.database.raw(sql)).rows.map((r) => r.uid);
-    }
-
-    // with local bumps
-    const fullCount = params.limit + params.offset;
-    const postsSQL = getPostsSQL(fullCount, 0, 'bumped');
-    const localBumpsSQL = pgFormat(`
-        with local_bumps as (
-          select post_id, min(created_at) as created_at from local_bumps where user_id = %L group by post_id
-        )
-        select b.post_id as uid, b.created_at as date
-        from
-          local_bumps b
-          join posts p on p.uid = b.post_id
-        where
-          ${sourceConditionSQL} and ${restrictionsSQL}
-        order by b.created_at desc
-        limit %L
-    `, viewerId, fullCount);
-
-    const [
-      { rows: postsData },
-      { rows: localBumpsData },
-    ] = await Promise.all([
-      this.database.raw(postsSQL),
-      this.database.raw(localBumpsSQL),
-    ]);
-
-    // merge these two sorted arrays
-    const result = [];
-
-    {
-      const idsCounted = new Set();
-      let i = 0, j = 0;
-
-      while (i < postsData.length && j < localBumpsData.length) {
-        if (postsData[i].date > localBumpsData[j].date) {
-          const { uid } = postsData[i];
-
-          if (!idsCounted.has(uid)) {
-            result.push(uid);
-            idsCounted.add(uid);
-          }
-
-          i++;
-        } else {
-          const { uid } = localBumpsData[j];
-
-          if (!idsCounted.has(uid)) {
-            result.push(uid);
-            idsCounted.add(uid);
-          }
-
-          j++;
-        }
-      }
-
-      while (i < postsData.length) {
-        const { uid } = postsData[i];
-
-        if (!idsCounted.has(uid)) {
-          result.push(uid);
-          idsCounted.add(uid);
-        }
-
-        i++;
-      }
-
-      while (j < localBumpsData.length) {
-        const { uid } = localBumpsData[j];
-
-        if (!idsCounted.has(uid)) {
-          result.push(uid);
-          idsCounted.add(uid);
-        }
-
-        j++;
-      }
-    }
-
-    return result.slice(params.offset, fullCount);
-  }
-
   async getTimelinesIntersectionPostIds(timelineId1, timelineId2) {
     const res1 = await this.database('posts').select('uid', 'updated_at').orderBy('bumped_at', 'desc').whereRaw('feed_ids && ?', [[timelineId1]])
     const postIds1 = res1.map((record) => {
@@ -522,236 +296,6 @@ const postsTrait = (superClass) => class extends superClass {
     const { rows } = await this.database.raw(sql);
     return rows.map((r) => r.uid);
   };
-
-  /**
-   * Returns array of objects with the following structure:
-   * {
-   *   post: <Post object>
-   *   destinations: <array of {id (feed UID), name (feed type), user (feed owner UID)}
-   *                 objects of posts' destination timelines>
-   *   attachments: <array of Attachment objects>
-   *   comments: <array of Comments objects>
-   *   omittedComments: <number>
-   *   likes: <array of liker's UIDs>
-   *   omittedLikes: <number>
-   * }
-   */
-  async getPostsWithStuffByIds(postsIds, viewerId = null, params = {}) {
-    if (_.isEmpty(postsIds)) {
-      return [];
-    }
-
-    params = {
-      foldComments:        true,
-      foldLikes:           true,
-      maxUnfoldedComments: 3,
-      maxUnfoldedLikes:    4,
-      visibleFoldedLikes:  3,
-      hiddenCommentTypes:  [],
-      ...params,
-    };
-
-    const uniqPostsIds = _.uniq(postsIds);
-
-    const postFields = _.without(Object.keys(POST_FIELDS), 'comments_count', 'likes_count', 'friendfeed_url').map((k) => pgFormat('p.%I', k));
-    const attFields = Object.keys(ATTACHMENT_FIELDS).map((k) => pgFormat('%I', k));
-    const commentFields = Object.keys(COMMENT_FIELDS).map((k) => pgFormat('%I', k));
-
-    const destinationsSQL = pgFormat(`
-      with posts as (
-        -- unwind all destination_feed_ids from posts
-        select distinct
-          p.uid,
-          unnest(p.destination_feed_ids) as feed_id
-        from 
-          posts p
-        where 
-          p.uid in (%L)
-      )
-      select
-        p.uid as post_id, f.uid as id, f.name, f.user_id as user
-      from 
-        feeds f join posts p on f.id = p.feed_id
-    `, uniqPostsIds);
-
-    const [
-      bannedUsersIds,
-      friendsIds,
-      postsData,
-      attData,
-      { rows: destData },
-    ] = await Promise.all([
-      viewerId ? this.getUserBansIds(viewerId) : [],
-      viewerId ? this.getUserFriendIds(viewerId) : [],
-      this.database.select('a.old_url as friendfeed_url', ...postFields).from('posts as p')
-        .leftJoin('archive_post_names as a', 'p.uid', 'a.post_id').whereIn('p.uid', uniqPostsIds),
-      this.database.select(...attFields).from('attachments').orderBy('ord', 'asc').orderBy('created_at', 'asc').whereIn('post_id', uniqPostsIds),
-      this.database.raw(destinationsSQL),
-    ]);
-
-    const nobodyIsBanned = bannedUsersIds.length === 0;
-
-    if (nobodyIsBanned) {
-      bannedUsersIds.push(unexistedUID);
-    }
-
-    if (friendsIds.length === 0) {
-      friendsIds.push(unexistedUID);
-    }
-
-    const allLikesSQL = pgFormat(`
-      select
-        post_id, user_id,
-        rank() over (partition by post_id order by
-          user_id in (%L) desc,
-          user_id in (%L) desc,
-          created_at desc,
-          id desc
-        ),
-        count(*) over (partition by post_id) 
-      from likes
-      where post_id in (%L) and user_id not in (%L)
-    `, [viewerId], friendsIds, uniqPostsIds, bannedUsersIds);
-
-    const foldLikesSql = params.foldLikes ? pgFormat(`where count <= %L or rank <= %L`, params.maxUnfoldedLikes, params.visibleFoldedLikes) : ``;
-    const likesSQL = `
-      with likes as (${allLikesSQL})
-      select post_id, array_agg(user_id) as likes, count from likes
-      ${foldLikesSql}
-      group by post_id, count 
-    `;
-
-    // Don't show comments that viewer don't want to see
-    let hideCommentsSQL = 'true';
-
-    if (params.hiddenCommentTypes.length > 0) {
-      if (params.hiddenCommentTypes.includes(Comment.HIDDEN_BANNED) && !nobodyIsBanned) {
-        hideCommentsSQL = pgFormat('user_id not in (%L)', bannedUsersIds);
-      }
-
-      const ht = params.hiddenCommentTypes.filter((t) => t !== Comment.HIDDEN_BANNED && t !== Comment.VISIBLE);
-
-      if (ht.length > 0) {
-        hideCommentsSQL += pgFormat(' and hide_type not in (%L)', ht);
-      }
-    }
-
-    const viewerIntId = viewerId ? await this._getUserIntIdByUUID(viewerId) : null;
-
-
-    const allCommentsSQL = pgFormat(`
-      select
-        ${commentFields.join(', ')}, id,
-        rank() over (partition by post_id order by created_at, id),
-        count(*) over (partition by post_id),
-        (select coalesce(count(*), 0) from comment_likes cl
-          where cl.comment_id = comments.id
-            and cl.user_id not in (select id from users where uid in (%L))
-        ) as c_likes,
-        (select true from comment_likes cl
-          where cl.comment_id = comments.id
-            and cl.user_id = %L
-        ) as has_own_like
-      from comments
-      where post_id in (%L) and (${hideCommentsSQL})
-    `, bannedUsersIds, viewerIntId, uniqPostsIds);
-
-    const foldCommentsSql = params.foldComments ? pgFormat(`where count <= %L or rank = 1 or rank = count`, params.maxUnfoldedComments) : ``;
-    const commentsSQL = `
-      with comments as (${allCommentsSQL})
-      select ${commentFields.join(', ')}, id, count, c_likes, has_own_like from comments
-      ${foldCommentsSql}
-      order by created_at, id
-    `;
-
-    const [
-      { rows: likesData },
-      { rows: commentsData },
-    ] = await Promise.all([
-      this.database.raw(likesSQL),
-      this.database.raw(commentsSQL),
-    ]);
-
-    const results = {};
-
-    const postsCommentLikes = await this.getLikesInfoForPosts(uniqPostsIds, viewerId);
-
-    for (const post of postsData) {
-      results[post.uid] = {
-        post:            initPostObject(post),
-        destinations:    [],
-        attachments:     [],
-        comments:        [],
-        omittedComments: 0,
-        likes:           [],
-        omittedLikes:    0,
-      };
-      results[post.uid].post.commentLikes = 0;
-      results[post.uid].post.ownCommentLikes = 0;
-      const commentLikesForPost = postsCommentLikes.find((el) => el.uid === post.uid);
-
-      if (commentLikesForPost) {
-        results[post.uid].post.commentLikes = parseInt(commentLikesForPost.post_c_likes_count);
-        results[post.uid].post.ownCommentLikes = parseInt(commentLikesForPost.own_c_likes_count);
-      }
-    }
-
-    for (const dest of destData) {
-      results[dest.post_id].destinations.push(_.omit(dest, 'post_id'));
-    }
-
-    for (const att of attData) {
-      results[att.post_id].attachments.push(initAttachmentObject(att));
-    }
-
-    for (const lk of likesData) {
-      results[lk.post_id].likes = lk.likes;
-      results[lk.post_id].omittedLikes = params.foldLikes ? lk.count - lk.likes.length : 0;
-    }
-
-    for (const comm of commentsData) {
-      if (!nobodyIsBanned && bannedUsersIds.includes(comm.user_id)) {
-        comm.user_id = null;
-        comm.hide_type = Comment.HIDDEN_BANNED;
-        comm.body = Comment.hiddenBody(Comment.HIDDEN_BANNED);
-        comm.c_likes = '0';
-        comm.has_own_like = null;
-      }
-
-      const comment = initCommentObject(comm);
-      comment.likes       = parseInt(comm.c_likes);
-      comment.hasOwnLike  = Boolean(comm.has_own_like);
-      results[comm.post_id].comments.push(comment);
-      results[comm.post_id].omittedComments = (params.foldComments && comm.count > params.maxUnfoldedComments) ? comm.count - 2 : 0;
-
-      if (params.foldComments && results[comm.post_id].omittedComments > 0) {
-        let omittedCLikes = results[comm.post_id].post.hasOwnProperty('omittedCommentLikes') ?
-          results[comm.post_id].post.omittedCommentLikes :
-          results[comm.post_id].post.commentLikes;
-
-        let omittedOwnCLikes = results[comm.post_id].post.hasOwnProperty('omittedOwnCommentLikes') ?
-          results[comm.post_id].post.omittedOwnCommentLikes :
-          results[comm.post_id].post.ownCommentLikes;
-
-        omittedCLikes -= comment.likes;
-        omittedOwnCLikes -= comment.hasOwnLike ? 1 : 0;
-        results[comm.post_id].post.omittedCommentLikes = omittedCLikes;
-        results[comm.post_id].post.omittedOwnCommentLikes = omittedOwnCLikes;
-      } else {
-        results[comm.post_id].post.omittedCommentLikes = 0;
-        results[comm.post_id].post.omittedOwnCommentLikes = 0;
-      }
-    }
-
-    for (const post of postsData) {
-      if (!results[post.uid].post.hasOwnProperty('omittedCommentLikes')) {
-        results[post.uid].post.omittedCommentLikes = 0;
-        results[post.uid].post.omittedOwnCommentLikes = 0;
-      }
-    }
-
-    return postsIds.map((id) => results[id] || null);
-  }
 
   // Insert record to 'archive_post_names' table for the test purposes.
   async setOldPostName(postId, oldName, oldUrl) {
@@ -878,7 +422,7 @@ const POST_COLUMNS_MAPPING = {
   isPropagable: (is_propagable) => {return is_propagable === '1'},
 }
 
-const POST_FIELDS = {
+export const POST_FIELDS = {
   uid:                  'id',
   created_at:           'createdAt',
   updated_at:           'updatedAt',
