+++ app/support/DbAdapter/utils.js
@@ -4,8 +4,6 @@ import pgFormat from 'pg-format';
 import { List } from '../open-lists';
 
 
-export const unexistedUID = '00000000-0000-0000-C000-000000000046';
-
 export function initObject(classDef, attrs, id, params) {
   return new classDef({ ...attrs, id, ...params });
 }
@@ -28,9 +26,6 @@ export function prepareModelPayload(payload, namesMapping, valuesMapping) {
 // We don't escape 'field' here because pgFormat escaping doesn't work properly with dot-joined
 // identifiers (as in 'table.field').
 
-// export const sqlIn = (field, list) => list.length === 0 ? 'false' : pgFormat(`${field} in (%L)`, list);
-// export const sqlNotIn = (field, list) => list.length === 0 ? 'true' : pgFormat(`${field} not in (%L)`, list);
-
 export function sqlIn(field, list) {
   list = List.from(list);
 
@@ -44,11 +39,11 @@ export function sqlIn(field, list) {
 }
 
 export function sqlNotIn(field, list) {
-  return sqlIn(field, List.difference(List.everything(), list));
+  return sqlIn(field, List.inverse(list));
 }
 
 export function sqlIntarrayIn(field, list) {
-  list = new List(list);
+  list = List.from(list);
 
   if (list.isEmpty()) {
     return 'false';
